Delineation of responsibilities (#1)
* Delineation of responsibilities

Introduces the notion of `ingest_api` and `client`.

`ingest_api` is responsible for formatting messages correctly for
different APIs such as humio-structured, humio-unstructured, hec, etc.,
as well as knowing how to construct headers and the path for those
specific APIs.

`client` is responsible for sending the messages formatted by
`ingest_api` to Humio.  Currently the only clients are a test client and
a Tesla client.  There are currently no plans to expand that.

* Update lib/client.ex

Co-authored-by: Daron Medina Caro <dairon.medina@gmail.com>

* Update lib/ingest_api.ex

Co-authored-by: Daron Medina Caro <dairon.medina@gmail.com>

Co-authored-by: Daron Medina Caro <dairon.medina@gmail.com>